Time: 1 ms (61.12%), Space: 43.8 MB (55.24%) - LeetHub

+++ 98-validate-binary-search-tree/98-validate-binary-search-tree.java
@@ -24,7 +24,7 @@ class Solution {
         if(root==null)
             return true;
         if(min<root.val && root.val<max){
-           return helper(root.left,min ,(long)root.val) && helper(root.right, (long)root.val,max);
+           return helper(root.left,min ,root.val) && helper(root.right, (long)root.val,max);
             
         }
